se actualizo y corrigio el login

+++ view/listar_usuario.php
@@ -1,5 +1,5 @@
 <div class="row">
-<?php if (count($dataToView["data"]) > 0): ?>
+	<?php if (count($dataToView["data"]) > 0): ?>
 		<table class="tabl">
 			<tr class="fila_tabla2">
 				<?php
@@ -18,7 +18,7 @@
 					$id = "id";
 					foreach ($campos as $key => $encabezado) {
 						if ($encabezado !== "" && $encabezado !== "ID") {
-							echo "<td>"; 
+							echo "<td>";
 							if ($key == "id_rol") {
 								echo $tabla["rol"];
 							} else {
@@ -28,18 +28,25 @@
 						} elseif ($encabezado == "ID") {
 							$id = $tabla[$key];
 						}
-
 					}
+					if (isset($_SESSION["rol"]) && ($_SESSION["rol"] == 2 || $_SESSION["rol"] == 1)) {
 					?>
+						<td>
+							<a href="index.php?controller=usuario&action=edit&id=<?php echo $id; ?>" class="btn btn-primary">✏️</a>
+						</td>
+						<?php
+						if ($_SESSION["rol"] == 1) {
+						?>
+						<td>
+							<a href="index.php?controller=usuario&action=confirmDelete&id=<?php echo $id; ?>" class="btn btn-danger">🗑️</a>
+						</td>
+					<?php
+						}
+					} ?>
 
-					<td>
-						<a href="index.php?controller=usuario&action=edit&id=<?php echo $id; ?>" class="btn btn-primary">✏️</a>
-					</td>
-					<td>
-						<a href="index.php?controller=usuario&action=confirmDelete&id=<?php echo $id; ?>" class="btn btn-danger">🗑️</a>
-					</td>
 				</tr>
 			<?php endforeach; ?>
+
 		</table>
 	<?php else: ?>
 		<div class="alert alert-info">
